UI + Bug fix
-Swiped cards were not being displayed properly in result table view
-Refactored result table view to be standard view controller with
tableView in VC
